AddGenreAndAuthorController

+++ BookStore.WebApi/AuthorOperation/Commands/DeleteAuthorCommands/DeleteAuthorCommand.cs
@@ -13,9 +13,11 @@ public class DeleteAuthorCommand
     }
     public async Task handleAsync(int id)
     {  
-        var author = await _context.Authors.Where(a => a.Id == id).FirstOrDefaultAsync();
+        var author = await _context.Authors.AsQueryable().Include(a => a.Books).Where(a => a.Id == id).FirstOrDefaultAsync();
         if(author == null)
             throw new Exception("yazar mevcut değil");
+        if(author.Books.Count() > 0) 
+            throw new Exception("yazar yayımda olan kitabı mevcut. once kitap'i sil gardasss");
         _context.Authors.Remove(author);
         await _context.SaveChangesAsync();
     }
