+++ BookStore.WebApi/GenreOperation/Commands/DeleteGenreCommands/DeleteGenreCommand.cs
@@ -15,8 +15,9 @@ public class DeleteGenreCommand
     }
     public async Task handleAsync(int id)
     {
-       var genre =  await  _context.Genres.FirstOrDefaultAsync(g => g.Id == id);
+       var genre =  await  _context.Genres.AsQueryable().Include(a => a.BookGenres).FirstOrDefaultAsync(g => g.Id == id);
        if(genre == null) throw new Exception("Kategori mevcut degil");
+       if(genre.BookGenres.Count() > 0) throw new Exception("Kategoriye ait kitaplar mevcut.Once Kitaplari siliniz ya da guncelleyiniz");
        _context.Genres.Remove(genre);
        await _context.SaveChangesAsync();
     }
